BF: read config from stdin instead of temporary file
as outlined in a comment by @bpoldrack in the code, git config -f - can
read from stdin. This is supported by git as its a unix convention.
Trial and error confirms that this also runs successfully in a Windows
CMD. By switching to a git config call with stdin as the input file, we
spare the need to write a temporary config file.
This fix could thus fix #6514, where a user error lead to a non-existent
TMP path and a subsequent failure to create the temporary config file.