Fix test failure with old annex
This test asserts that the push call will fail, because of read-only
mode of the ORA remote. This is what happens, but with older annex we
get an `IncompleteResultsError` rather than a `CommandError`. Not sure
what changed for this to happen now, but in any case:
`IncompleteResultsError` would be more desireable to begin with as we
get an actual error result rather than a crash + if there's something to
change to get this with all annex versions, it's in `push` or at a more
generic layer when calling annex. The test, however, only is about
testing that annex-copy and by extension datalad-push with the ORA
remote fails under these circumstances.
Hence, accept both ways of failing from within this test. The difference
warrants an investigation of `push`/`Runner` or whatever is causing the
difference here, though.

Closes #7157