BF: Prevent doomed annex call on untracked files
Fixes #7032. When datalad status --annex is called on an untracked
file specifically, it will invoke an annex query that we already know
will fail. This commit adds a check if any specifically added file as a
path is untracked, and if so, prevent the annex query from being run on
it.