record runner-owning thread
This commit adds code to record the thread that currently
"owns" or "uses" a ThreadedRunner-instance. This allows to
improve the message of the re-entered-RuntimeError.

The commit robustifies the generator by ensuring that the
exhausted state is actually trivial, i.e. calling send() on
an exhausted generator will only use minimal information
that is internal to the generator.

In addition it tests for None-message in the first send()-call.