BF: Fix invalid absolute submodule paths when cloning relative paths
This change reincarnates parts of a changeset originally proposed in
https://github.com/datalad/datalad/pull/4026, that didn't make it in due
to a failing Windows test that Ivnever saw (CI logs weren't preserved),
but that I suspect to originate
in #7180 (Git stitching together posix and windows paths into a
non-functional URL). Thus, this change sits on top of #7181, which fixes
these URLs to be fully posix.
Based on those fixes, this change wrangles absolute URLs originating
from relative paths back into relative paths, to keep them functional.
This change also enables an old test for this problem, marked as a known
failure.

Fixes #3538