Ensure to-be-exported archive is represented as a string
This is a bandaid, because methods such as repo.add used down the line
are incapable of handling Pathlib objects. A pathlib object instead of a
string would also torpedo the progressbar generation, as the archive name
is passed as a label (which is expected to be a string). The latter case
is also a bandaid, especially given that get_progress_bar is outdated
and alsmost nowhere used but there and on the list for an overhaul
(see gh-6721)