do not store runner in WitlessRunner & clean up
This commit removes the storing of ThreadedRunner
instances in WitlessRunner. The stored runner is
not used and it is not useful since multiple threads
may enter WitlessRunner.run() and overwrite the
stored runner in unpredictable ways

The commit also improves the description of the
`cwd`-parameters and their behavior.

The commit fixes a problem with path reporting in
CommandErrors which could appear if a non None
`cwd` was provided to `WitlessRunner.run()`. I.e.
the default `cwd` set in the constructor would
have been reported instead of the actual `cwd`,
provided when `WitlessRunner.run()` was called.

The tests are extended to check for multi-thread
support of `WitlessRunner.run()` in the
non-generator mode.