fix and error in process exit checking
This commit fixes an error where a process exit was
falsly assumed when poll() returned None. That gets
rid of the "lost process"-state.

Also adapt the tests to the new synchronization
in generator runners, i.e. block a new run on an
active generator runner until the existing generator
is exhausted.