DOC: Use standard key name 'error_message' in result record design docs
As elaborated in detail by @mslw in #6620, the common keyname for an error message
in a result record is error_message. The Design document on result records
however spells this key as error-messages, not as error_message as understood
by generic_result_renderer(). This commit changes the design document to use
error_message. Fixes #6620