TST: Mark some tests as known fails on github CI windows
Ref: #7184, #7185, #7197

Need to do that since there is no human force available to investigate
and address them currently and we might this way miss more failures on
datalad/git-annex CI setup.

Note: that I marked them as known only for github CI, so they might
resurface on appveyor whenever a new release of git-annex comes out.
Then either they would need to be addressed or marked as generally known
to fail on windows