NF: make extra_remote_settings configurable in git config
This makes the extra_remote_settings attribute for GitHub-like create_sibling
targets configurable on a per target basis. It introduces a new configuration
section

	"datalad.create-sibling-ghlike.extra-remote-settings.NETLOC.KEY=VALUE"

and applies all specified "KEY=VALUE" configuration to any sibling created via
create_sibling_* on the GitHub-like instance at "NETLOC". One usecase would be
to override an annex-ignore setting, which is otherwise unconditionally set.
E.g. if we host a gitea server which does support git-annex at example.com,
we could set:

	[datalad "create-sibling-ghlike.extra-remote-settings.example.com"]
		annex-ignore = false

in e.g. ~/.gitconfig to override the default of not using git-annex with gitea.