BF: Don't change ui.backend of test process
The `main()` function for special remotes sets the `ui.backend`.
Intended target is a dedicated special remote process this is expecting
to run it. However, when testing this very `main` function, the python
context this would change the backend for, is the entire test process.
This subsequently led to failures in other tests, because underlying
code would start to behave interactively again.