re-add annex file-URL support
This commit cleans up the path handling more. It improves
conversion between local os-specific paths and URL-path
components.

This commit brings back support for the git-annex on Windows
convention of putting the anchor of a path, e.g. 'C:', into
the network location instead of doing it as defined in
Appendix E of RFC 8089.