clone auto-reconfigures ORA remotes only locally
`clone` does an auto reconfiguration of ORA special remotes, if the
dataset was cloned from a RIA store, an ORA remote failed to autoenable
and clone finds that the special remote's uuid in the store it just
cloned from. The clone URL would then be configured to be used by the
ORA special remote. However, until now that new URL was passed
to `annex enableremote` and thereby committed in the git-annex branch.

This could lead to a back and forth of changing committed URLs by
different users who may have access to that store via different methods
(SSH/HTTP). This patch changes the reconfiguration to only happen
locally in .git/config, since ORA does now respect and gives precedence
to local configs.