BF: Local reconfig of ORA failed during clone
ORA's consideration of local configs for overriding stored configs
relied on finding the remote name used in git config. However, during
`initremote` on a fresh clone, this name can't be determined yet and ORA
failed to go with the default name instead, which was used by the RIA
clone patch.