Add private (annex) option to create
This adds a new parameter, "private", to the create command. When set,
it will set the annex.private configuration option before calling git
annex init. When the option is set, the created dataset won't record
information about itself in the git-annex branch (but the branch will
still exist and record e.g. information about other repositories).

This can be useful for creating temporary datasets which will be
discarded after being published, or for privacy reasons.