synchronise BatchedCommand cleanup
This commit adds a lock to synchronise the
execution of BatchedCommand.cleanup() with
the executions of __call__ on BatchedCommand
instances. That means that no instance is
cleaned while executing a command, i.e.
sending a request and waiting for a
resppnse. The cost for that is that
cleanup cannot run during execution of any
BatchedCommand.__call__ and therefore
no BatchedCommand.__init__ will
terminate during this time, so creation
of new instances can only happen if
no BatchedCommand is currently executing.