split up global BatchedCommand.cleanup lock
This commit removes the global BatchedCommand
cleanup-lock in favor of inidividual locks in
the BatchedCommand instances. This prevents the
instances from serializing external calls, but
requires a synchronisation point that includes
all existing instances, when a new instance
is created.

In addition this commit renames the global
instance dictionary in order to reflect that it
contains all instances that have a runner
associated.

Work in the next commits should contain:
add a limit to the cleanup duration, use a
WeakSet instead of a WeakDictionary for the
global instance collection.