BF: Fix invalid version comparison
Specifying 10.20221212+git18 as requirement doesn't work, when the
reported version of an annex build is using a different scheme. It was
used because the git-annex version required isn't released yet. See
issue 7248.

However, we know the scheme of annex versions (it's a date after the
dot), therefore setting the required version to one day after the
previous release does the trick with respect to released annex versions.

Closes 7248