Add tests for private mode in clone & create
This should also cover get, because all that's relevant happens in
create.

Note that checking the presence of "annex.private" in config is not
sufficient. To make a repository private, this option must be in place
before git annex init ts called. So to be sure, we also check effects
of this config, i.e. presence of "journal-private" & absence of new
entries in git-annex::uuid.log.