BF: path from .gitmodules could not be used with source candidate template
The `path` property was treated as a special case in
`GitRepo.get_submodules_` presumably to not yield redundant information,
because what is returned is a dict where the path is the key and the
other properties from `.gitmodules` make up the value (another dict),
prefixed with `gitmodule_`.

However, actual usage from `get` via `subdatasets` suggests, that the
easiest is to just yield this "redundant" record. Once as the reported
property exactly like `url`, `datalad-id` and whatever else, and in
addition let the path remain the key. Throughout that chain the latter
gets possibly turned into an absolute path anyway (which is not what we
want to report here - it's just for internal use).

Closes #7274