RF: WitlessRunner do not raise CommandError if exit non-0 and exception_on_error=False
Original behavior to raise that CommandError after process exits was
introduced in 9551b91929f6ee8e48a92fb343cdfdfbc05cef22 (0.13.0rc1~120^2~1)
whenever exception_on_error was added later on in 9f066efa6a0b316a133b6d6d0d4e4c2e1b856ce0
(0.16.0^2~123^2~7).

It is not clear to me (question to @mih) why it was `get` with defaulting to
`True`, so I have changed it to `None`.  I have then adjusted the logic to actually
react to when there was nothing (we got None stored or got as default) really to
match comment before. And made it to raise CommandError for non-0 only if exception_on_error was True.