Reduce test environment complexity by redirecting HOME
Since Git 2.32 we are still creating a dedicated temporary HOME
directory for running tests, but no longer actually enable it to be used
beyond the global git config (i.e., `.gitconfig`).

This change rectifies that and actually reconfigures the environment
to use this temporary HOME.

This should more effectively avoid "caryy-over effects" between test
runs, and make debugging easier (because of more similar conditions with
the CI systems that are destroyed entirely after each run).

Without this change, tests could still trigger changes in a real user's
`$HOME/.config` or `$HOME/.cache` that are persistent across test runs.

A simple test is included to trigger breakage in case someone is making
modifications to this setup.

Closes #7297