BF: Don't skip subds underneath paths in reckless & recursive drop
This fixes #7013.
If a drop --recursive --what all --reckless kill was provided with paths
to directories, it silently skipped killing the contained subdatasets.
The problem arises when the internal resolving of paths to datasets
determines that the provided directory path belongs to the superdataset.
Instead of then recursing into the subdirectory's potential subdatasets,
drop restrained the scope of what was to be dropped to only
'filecontent' in order to limit the drop from the determined
superdataset to the provided path as an internal safety mechanism:

if paths is not None and paths != [ds.pathobj] and what == 'all':
    # so we have paths constraints that prevent dropping the full dataset
    lgr.debug('Only dropping file content for given paths in %s, '
              'allthough instruction was to drop %s', ds, what)
    what = 'filecontent'

This change lifts this safety mechanism partially: Whenever reckless is
set to 'kill', and the provided path does not match the currently
investigated dataset, it adds a subdataset call constrained to the path
in question. If this returns a dataset, we know that this dataset is
meant to be dropped as well, as recursive (which has to be set whenever
reckless==kill) must be set, too.