BF: GitRepo.merge do not allow merging unrelated unconditionally
As @bpoldrack identified in
https://github.com/datalad/datalad/issues/7291#issuecomment-1456147953 the
629e575628ba8b53bf073392837c9e4b49444838 which was part of the #4650 concerning
introduction of minimal git version to be 2.19.1, and not removing that
condition was was combined with argument to the merge function. That commit was
released within good old 0.13.4.

This change merely reintroduces "if" statement BUT unfortunately that
allow_unrelated nowhere exposed in the API really, and used only once with
allow_unrelated=True in some single test. So I expect some tests to fail here
and most likely us needing to introduce similar option in "update" to finalize
this PR.