Adopt call_annex_success helper from next
Originally implemented in datalad-next by @mih and also proposed in #6316,
this change migrates a helper from datalad-next that wraps annexrepo.call_annex()
with a convenience layer to return True or False depending on the success of the call.
I found at least one usage in annexrepo where this helper is useful, and I suspect
there are more. Fixes #6316