Transition s3 downloaders to boto3
This commit replaces the usage of boto (which has been deprecated)
with boto3 in the s3 downloaders module.

Instead of boto bucket and key objects, the functions in this module
will now operate primarily on a boto3 client object ("low-level
interface to AWS whose method map close to 1:1 with service
APIs"). Although boto3 has a high-level resource object for a Bucket,
it behaves differently than the old boto bucket, and I found working
with a client much more suitable for the kind of operations done by
the downloaders code (setting credentials, asking about headers,
etc.).

The commit also allows boto3 to use credentials coming from its own
configuration mechanism (`AWS_*` environment variables,
`~/.aws/credentials`, `~/.aws/config`, `/etc/boto.cfg` and `~/.boto`)
allowing AWS users to rely on their existing configuration. The
preference order is DataLad credentials, boto credentials, anonymous.

Some unused code pieces (now incompatible with the rest of the file)
are left over, and some new TODOs for completing the transition are
introduced. These will be addressed by subsequent commits. But the
current code works for simple (authenticated & anonymous) download
scenarios already.

Note: The S3Authenticator.authenticate docstring said that it
authenticates to the specified bucket. This behaviour was maintained
for compatibility - we effectively check bucket permissions, as the
old code did. But the transition to boto3 client object makes it
apparent that authenticating to a bucket is not really what we care
about if all we want is to download an object. Authentication is done
to S3, and then object permissions are decisive for downloading. For
example, there can be a public file (public read permission) in a
bucket that doesn't block public access, but does not grant public
permission to list objects (easy and reasonable setup in S3) - in
which case non-owner bucket interaction will end in 403 although the
object could be retrieved.