Retire S3Downloader's _get_key, _get_key_via_get methods
The _get_key / _get_key_via_get approach was used to gain additional
insight from the get response if the first (head) attempt fails. I am
wary of the fact that fixes usually accumulate in the code for good
reasons, but I feel that these can be retired without massive
consequences.

Admittedly, we currently only perform head requests (head_object in
this context), which have no response body and therefore return just a
generic 404 Not Found or 403 Forbidden code if things go wrong. But I
think this is workable already. However, if the get query is deemed
necessary, the commit can be reverted, and the logic rebuilt based on
get_object, which might be more convenient to work with.