Remove lookup of s3 object VersionId
This was introduced as part of the boto to boto3 rewrite, but if the
code comment about NDA still holds true (I did not verify), it is
redundant. We simply won't add VersionId to download parameters unless
it was contained in the URL, no unsetting necessary. Original comment
introduced in 34ba22a is left for posterity.