Use boto3 retry features
Boto3 itself is capable of retrying calls to AWS. This commit removes
downloader's usage of our custom try_multiple_dec_s3 and instructs
boto3 to use its standard retry mode (default is legacy). The standard
mode will do a default of 3 maximum retry attempts with an exponential
backoff time.

https://boto3.amazonaws.com/v1/documentation/api/latest/guide/retries.html

As it stands, the mode setting will override user aws config (if any
is present at all), but I think it's fair. Alternatively we could do
the same thing as for credentials, i.e. create a session first and see
if it picks up any retry mode. Possible modes are: legacy, standard,
and adaptive (the latter is described as experimental).