Migrate most of s3 support code to boto3
It seems that most of support/s3.py documents the process of creating
test buckets - these functions are not imported elsewhere. Usage of
some functions had been removed when transitioning s3 downloader to
boto3. Judging by git grep, now only get_versioned_url and
add_version_to_url are used elsewhere (the former in local/addurls.py,
the latter in tests).

Functions that aren't usied locally (to generate test buckets) or
imported elsewhere are now removed. Generation of test buckets is
migrated to boto3, and based on on high-level resources (which are
different from low-level clients used in the downloaders code, but
more fit for purpose here).

In manual testing, I had to rename the buckets (otherwise, I got 403),
but after doing so I was able to run all the bucket-generating code
under my aws account.

The get_versioned_url function is yet to be reworked and isn't
functional as of this commit - but because it has a lot of logic, the
task is left for a separate commit.