TST(BK): extend test to demonstrate https://github.com/datalad/datalad/issues/7351
Although the file is about .repo.save and not ds.save, there is already a good
number of other tests which pretty much test ds.save.  It is also the only
place which already uses that convoluted scenario so I decided it would be good
place for an addition of another testing thus not introducing much of extra run
time as if I was to create an independent new test for this problem.