Fix Providers caching of config files
Providers.from_config_files() stores the result of
`get_dataset_root("")` and delivers the previous resut, if not called
with `reload` and the dataset root is the same.
However, what effectively is stored when CWD is a dataset's root is the
empty string - so, effectively `.`. But this would be the same for any
dataset root we are in, therefore this lazy loading would not recognize,
that the dataset changed. This lead to side-effects of the order of test
execution, when a test provided new config files. Those files would be
ignored, when providers were read from within a different dataset
before.

Hence, use the absolute path to store and compare.

Closes #7347