Add datalad.ui.interactive config and non-interactive backend for special remotes
This patch introduces the config `datalad.ui.interactive` in order to
let users decide whether or not to run in interactive mode.
The config defaults to the former detection, except that this detection
is additionally safeguarded - any exception during that detection will
lead to non-interactive mode. In addition, the result of the detection
will be stored in `DATALAD_UI_INTERACTIVE`, thus passing down the result
to possible subprocesses rather than having them running their own
detection. Ultimately, this is about `ui`'s ability to talk to the
terminal via `getpass` and the detection does not work from within
subprocesses.

Closes #7349

Furthermore, this adds a non-interactive UI backend for annex special
remotes, which previously always assumed to be in interactive mode, and
adds the respective capacity for the UI_Switcher.

Closes #7345