ENH(TST): add DATALAD_TESTS_CREDENTIALS=system for known credentials to be used during tests
We hardcoded to set it to a plain text keyring in
d8d538ae77ee8740ee2ca44e2b33f99b73a4ec3c of
https://github.com/datalad/datalad/pull/7209 to completely disable possible
attempt of keyring to interact with user e.g. to establish a new keyring.
That was done to address #6623 which came up during debian package building.

But the need for older behavior was realized while approaching #7340
where interactions with s3 were replaced to use boto3 but to test I needed
credentials, but we were just skipping the tests since credentials were not
present in that fake keyring.

With this patch I could run

  DATALAD_TESTS_CREDENTIALS=system python3 -m pytest -s -v datalad/downloaders/tests/test_s3.py

and see tests to interact with S3.  Test relating to NDA were failing
since most likely NDA broke the "flow" of how to access data there.