Finish private mode support
This introduces a version kludge indicating whether private mode is
supported by git-annex and adjusts tests for clone and create to account
for it.
Furthermore this patch makes `create` and `clone` fail if private mode
was requested by command parameter but isn't supported by installed
git-annex. This seems preferrable over a fallback solution like `git
annex dead here`, since the behavior is not the same and whether such an
approach suits the usecase is up to the user to decide.

There's one deviation from failing early when private mode isn't
supported: This is when the setting comes from the
`datalad.clone.reckless` setting in a superdataset. In this case we fail
only when actually trying to set this. The rationale for this is:
At the point in time we read such a config, we don't even know whether
we are creating/cloning an annex repo to begin with (could be plain git).
To fail when we could succeed, while the instruction for private mode
only comes from a general setting rather than by telling the executed
command directly seems wrong. Failing to `clone --reckless private` a
git repo seems OK, because it can be rectified by the command call that
doesn't need the option to begin with. But `clone -d.` a git repo while
`datalad.clone.reckless` is set to private in the superdataset seems
different, because a user would need to change the general setup for a
config that isn't even supposed to have an effect on the dataset in
question.
It may arguably be rare to ever run into this situation, but we should
be aware, that users may have good reasons for different execution
environments (incl. different annex versions), while working on the same
datasets on the filesystem (where the config for the superdataset is
likely to come from).