BF: git-annex now would return json record for trying to add a nonexisting file
Prior now we would have produced a fake json record. I have decided to pass new one
through, which would have its own error-messages. But our fake one also had a "note", so I
decided to add it for consistency (unless git-annex would start adding it).
We use that 'note' internally and seems only for comparison to 'not found'.
Now that ther is 'message-id', in principle, we could uniformly fake and use that one instead.