Discontinue ConfigManager abuse for Git identity warning
DataLad warns about an undefined Git identity (user.name|email).
It does that whether or not any Git identity is actually necessary
or even anyhow considered. The `ConfigManager` constructor is
(ab)used to trigger this warning, and a `ConfigManager` class
member is used to prevent repeated warnings.

This changeset maintains the exact same behavior (warn on `datalad
import`), but move the associated code out of the `ConfigManager`
class. This class is not involved in any Git repository
manipulations. The warning is trigger in `datalad/__init__.py`
directly, hence no protection against repeated execution is needed
(Python does it automatically).