BF+RF+TST: invoke git/annex with core.quotepath=false, unquote log for recent annex, fix+test unannex
Not quoting filenames is not default git behavior and might have some ramifications!

https://github.com/datalad/datalad/pull/7372 has more discussion on what inspired these changes.

"unannex" was completely broken in case of filenames with spaces, so this
commit fixes it as well while going through interfaces where we still do not
use --json.  Filed
https://git-annex.branchable.com/todo/--json_for_unannex__and_ideally_any_other_command_/
which was said to be already addressed.

Also added explicit testing in that added test for get_annexed_files which
uses find and which does not quote ATM.