BF: make addurls tollerate the case that not all rows have all metadata fields
Use case is the https://gpt4all.io/models/models.json  where two keys are used
only in a few records.  We can exclude them for all but I thought it would then
would effectively loose metadata.  An alternative is to provide an alternative
value e.g. "n/a" so may be worth an option?