Adjust dataset and path separator stubs for GitLabs naming rules
This change introduces two configurable stubs: project as an alternative
to _repo_ (as a name for collection and hierarchy layouts root or subdataset
projects) and - as a path separator for flat and collection subdataset names.
The defaults comply to GitLabs naming rules for project paths, introduced in
version 14.9:
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/80055

They remain configurable via dedicated dataset-level configurations to allow
overrides, for example to prevent name clashes with existing (sub)datasets
or if path restrictions on gitlabs side are lifted.
The configurations are:

* datalad.gitlab-default-projectname
* datalad.gitlab-default-pathseparator