Fix for broken hierarchy layout
This commit applies a small change to remove the
discrepancy between the docs description of the
hierarchy layout and the actual implementation.
The result is that each dataset name in a nested
structure, including the top-level dataset, will
map onto a gitlab group, and the associated dataset
repositories will be created as a projects inside
these nested groups.

An additional requirement is that a user needs to
create a top-level group via the Gitlab UI (this
cannot be done via the API) and this group needs to
be provided as the base of the value in the
--project argument, e.g. 'newgroup/mydatasets'.