BF: in hierarchy layout, remove subdirectories from subdataset paths
This is a fix for #7411.
In a dataset hierarchy with subdatasets in subdirectories, the
hierarchy layout of create-sibling-gitlab will keep path separators
in the project path. This is problematic, because for GitLab, this
makes it look like the subdirectory needs to be a pre-existing group.
For example:

[DS~0] /home/me/dl-101/DataLad-101
├── recordings/
│   └── [DS~1] longnow/

will create a project path like this:
<gitlab-instance>/<gitlab-group>/recordings/longnow/project

DataLad will attempt to create the group 'longnow', but won't consider
creating 'recordings'. Thus, the action will fail due to a missing
parent level group that was never meant to be created.

With this change, the project path becomes
<gitlab-instance>/<gitlab-group>/longnow/project

I.e., subdirectories are stripped.