work around a typing issue
This commit ensures locally, i.e. independent
of previous control flow, that containment is
tested against dictionaries and not against
None-values.

The type checker does not determine that a
non-None original environment will lead to
a non-None adjusted environment, when
Runner._get_adjusted_env is called. We help
it with an `assert` that states this fact.