Convert boto3's 403 into AccessDeniedError
There are two places where AWS can respond with 403 access denied,
causing botocore.exceptions.ClientError:

1. When the S3Authenticator calls head_bucket (note that this would be
a false positive for buckets that have public objects but don't allow
listing; however it follows the logic of "authenticating to a bucket"
that was present in boto2 code).

2. When the S3Downloader calls head_object

These ClientErrors are now caught and raised as DataLad's
AccessDeniedError.

IMHO this leads to less informative error messages, but
AccessDeniedError seems to be expected in other places of the
codebase. The ClientError message would say, e.g., "An error
occurred (403) when calling the HeadObject operation: Forbidden",
indicating whether it came from HeadBucket or HeadObject. To preserve
the error message, at least in debug output, I am using the original
error as the message for AccessDeniedError.

What is certain is that raising AccessDeniedError was necessary to
make test_deny_access pass when "doesn't matter" (used by the test as
a download url) was replaced with an url to an actual S3 bucket or
object that returned 403. The test as it is still fails, because boto3
now tries to validate the url format and raises a different error (in
any case, not equivalent to access denied), which needs to be
addressed in the test itself, with next commit.