RF: remove travis CI setup + badge from README.md
We are using paid travis CI, and funds for that line are about to end.
Moreover @mih is also paying for appveyor.  So I think it would be reasonable
to remove travis. The tricky part would be to get coverage parity since travis
has AFAIK more extended battery of matrix runs to excercise different
deployment scenarios ATM.