BF: fix time_diff* and time_remove benchmarks to account for long RFed revision arg etc
Only with fresh asv 0.6.0 which started to exit with exit code 2 if any
benchmark failed, we finally had to pay attention to those long failing
benchmarks. We have removed the option 'revision' so long ago but first didn't
mention and then just ignored. For that and time_remove I also made it into
dataset method invocation since otherwise it was not finding the dataset any
longer given the path to it (didn't investigate deeper).