Robustify update target detection for adjusted mode datasets
The true cause of the problem is not well understood.
https://github.com/datalad/datalad/issues/7507 documents that I was not
able to capture the breakage in a test, despite approaching it from
different angles.

Only the original reproducer documents that this patch fixes the
problem. However, the involved datasets are too have to be used in an
already overloaded test battery. Hence this fix is proposed without a
corresponding test.

Closes #7507