Adjust some ORA remote's ssh commands based on remote OS
The ORA remote (more specifically, its `SSHRemoteIO.ensure_writeable`
context manager) runs the `stat` command on the remote end to check
file permissions. The command format and its output differ between
macOS and Linux.

Previously, the format was chosen based on the *local* OS, which makes
no sense for a command that is executed on the *remote* end. This led
to situations where it was impossible to push data from Mac to Linux
because the stat command errored out, and permissions were left
incorrect, when moving a file from `transfer` to its desired location.

With this change, the ORA remote will run `uname -S` on the remote end
(of an SSH connection) to determine where it is operating.

To the best of my knowledge (https://en.wikipedia.org/wiki/Uname),
checking for "Darwin" should be good.