Add remote uname as a lazy property of ORA remote's SSH IO
This adds a property `remote_uname`, with lazy resolution (by running
`uname -s` on the remote end), to the ORA remote's `SHRemoteIO` class.
Thus, the SSH commands can operate with an awareness of the remote OS,
if needed.

One situation when this is needed is the class's `ensure_writable`,
which needs to run `stat` with an OS-specific command (and output)
format. Having access to the property avoids re-running `uname` for
every file that needs to be touched, and having the property lazily
resolved avoids running `uname` for operations which don't need it.

Note that although `uname` is UNIX-specific, the class docstring
already says that "It doesn't even think about a windows server".