Update GitHub actions to use Python 3.8
Python 3.7 is long past its EOL now. We still officially support
it (python_requires>=3.7 in setup.py) but I feel a change there is
overdue, too.

The test_extensions workflow was updated in 3378d713b and test_macos
in 7ccd7b043 as they started showing problems. This commit changes the
remaining GitHub workflows which use the setup-python action to also
use 3.8.

Related to #7400