Replace Python 3.7 with 3.11 on Appveyor
This changes the tests labelled as "Test alternative Python version"
from Python 3.7 to 3.11.

A move to 3.8 would sound logical, but... if I'm seeing correctly, the
only difference between Ubu20P37a & Ubu20P37b versus Ubu20core and
Ubu20a1 is that the former had "PY: 3.7", while the latter did not set
PY at all. And if I'm reading the logs of the latest Appveyor run
correctly, Ubu20core & Ubu20a1 run on Python 3.8.10. I cound not find
confirmation that this is the default, but it seems to be.

An alternative would be to set PY for Ubu20core & Ubu20a1 to the
newest supported (3.11) and use the other two to test oldest
supported (3.8). But maybe appveyor tends to stay on the lowest
non-EOL version when PY is not specified, in which case the current
setup would be preferred (just a guess - couldn't find confirmation)?

Related to #7400