RF: centralize logic for considering an extension into iter_extensions
It will be guided by the same datalad.extensions.load ATM,
although I think we should refactor it in the long(er) term into
datalad.extensions.enable  which would be more descriptive of the
purpose.

IMHO we should either consider entire extension or not consider it at all.
ATM we have diverging logic which leads to confusion as "I have installed extension X
and it works, but I installed extension Y and its effects are not there"